+++ components/PageLayout/Header.js
@@ -16,10 +16,12 @@ const Header = React.createClass({
             permissions: PropTypes.array
         })),
         location: PropTypes.object,
+        buttonsRaw: PropTypes.node,
         breadcrumbsRemoveSlashes: PropTypes.number
     },
     getDefaultProps() {
         return {
+            buttonsRaw: [],
             buttons: [],
             location: undefined,
             breadcrumbsRemoveSlashes: 1
@@ -29,7 +31,7 @@ const Header = React.createClass({
         checkPermission: PropTypes.func
     },
     render() {
-        let { text, buttons, location, breadcrumbsRemoveSlashes } = this.props;
+        let { text, buttons, buttonsRaw, location, breadcrumbsRemoveSlashes } = this.props;
         let breadcrumbsString = '';
         if (location) breadcrumbsString = getBreadcrumbsStringFromPathString(location.pathname, breadcrumbsRemoveSlashes);
 
@@ -40,6 +42,11 @@ const Header = React.createClass({
                     {text}
                 </h1>
                 <div className={classnames('pull-xs-right', style.buttonsWrap)}>
+                    {buttonsRaw.map((btn, i) => {
+                        return (
+                            <div key={i} className={style.buttonWrap}>{btn}</div>
+                        );
+                    })}
                     {buttons.map((btn, i) => {
                         if (btn.permissions && !this.context.checkPermission(btn.permissions)) {
                             return <div />;
@@ -50,8 +57,7 @@ const Header = React.createClass({
                                 <div key={i} className={style.buttonWrap}>
                                     {(i > 0) && <span className='w20' />}
                                     <StandardButton
-                                      className='secondaryButton'
-                                      disabledClassName='disabledsecondaryButton'
+                                      styleType='secondaryDark'
                                       onClick={btn.onClick}
                                       disabled={btn.disabled === true}
                                       label={btn.text} />
@@ -62,8 +68,7 @@ const Header = React.createClass({
                                 <Link key={i} to={btn.href} className={style.buttonWrap}>
                                     {(i > 0) && <span className='w20' />}
                                     <StandardButton
-                                      className='secondaryButton'
-                                      disabledClassName='disabledsecondaryButton'
+                                      styleType='secondaryDark'
                                       disabled={btn.disabled === true}
                                       label={btn.text} />
                                 </Link>