+++ components/Popup/Header.js
@@ -5,7 +5,7 @@ import styles from './styles.css';
 const Header = ({
     className,
     text,
-    hasCloseBtn,
+    closeIcon,
     closePopup
 }) => {
     // Additional outer div is used to fix positioning issue in Firefox. Do not remove it.
@@ -13,21 +13,21 @@ const Header = ({
         <div className={classnames(styles.popupHeader, className)}>
             <div>
                 <span className={styles.headerText}>{text}</span>
-                {hasCloseBtn && <div className={styles.closeBtn} onClick={closePopup} />}
+                {closeIcon && <div className={styles.closeBtn} onClick={closePopup} />}
             </div>
         </div>
     );
 };
 
+Header.defaultProps = {
+    closeIcon: true
+};
+
 Header.propTypes = {
     className: PropTypes.string,
     text: PropTypes.string,
-    hasCloseBtn: PropTypes.bool,
+    closeIcon: PropTypes.bool,
     closePopup: PropTypes.func
-};
-
-Header.defaultProps = {
-    hasCloseBtn: true
 };
 
 export default Header;