+++ components/Popup/index.js
@@ -1,15 +1,26 @@
 import React, { Component, PropTypes } from 'react';
-import RenderToLayer from './RenderToLayer';
 import classnames from 'classnames';
+import debounce from 'lodash.debounce';
+import RenderToLayer from './RenderToLayer';
 import Header from './Header.js';
 import Footer from './Footer.js';
+import { POPUP_MIN_OFFSETS, POPUP_HEADER_HEIGHT, POPUP_FOOTER_HEIGHT } from './config';
+
 import styles from './styles.css';
 
 class PopupInternal extends Component {
     constructor() {
         super();
-
+        this.state = {
+            contentMaxHeight: ''
+        };
+        this.handleWindowResize = debounce(this.handleWindowResize.bind(this), 100);
+        this.updateContentMaxHeight = this.updateContentMaxHeight.bind(this);
         this.handleEsc = this.handleEsc.bind(this);
+    }
+
+    componentWillMount() {
+        window.addEventListener('resize', this.handleWindowResize);
     }
 
     componentDidMount() {
@@ -18,14 +29,26 @@ class PopupInternal extends Component {
         if (closeOnEsc) {
             document.addEventListener('keydown', this.handleEsc);
         }
+        this.updateContentMaxHeight();
     }
 
     componentWillUnmount() {
         document.removeEventListener('keydown', this.handleEsc);
+        window.removeEventListener('resize', this.handleWindowResize);
     }
 
-    handleEsc(e) {
-        const { keyCode } = e;
+    handleWindowResize() {
+        this.updateContentMaxHeight();
+    }
+
+    updateContentMaxHeight() {
+        const contentMaxHeight = window.innerHeight - POPUP_MIN_OFFSETS - POPUP_HEADER_HEIGHT - POPUP_FOOTER_HEIGHT;
+        this.setState({
+            contentMaxHeight: `${contentMaxHeight}px`
+        });
+    }
+
+    handleEsc({ keyCode }) {
         const { closePopup } = this.props;
 
         if (keyCode === 27) {
@@ -49,8 +72,8 @@ class PopupInternal extends Component {
             <div className={styles.modalContainer}>
                 { hasOverlay && <div className={styles.modalOverlay} onClick={closeOnOverlayClick ? closePopup : null} /> }
                 <div className={classnames(styles.popupContainer, className)}>
-                    { header && <Header className={header.className} text={header.text} closePopup={closePopup} /> }
-                    <div className={classnames(styles.popupContent, contentClassName)}>
+                    { header && <Header className={header.className} text={header.text} closePopup={closePopup} closeIcon={header.closeIcon} /> }
+                    <div style={{maxHeight: this.state.contentMaxHeight}} className={classnames(styles.popupContent, contentClassName)}>
                         { children }
                     </div>
                     { footer && <Footer className={footer.className} actionButtons={footer.actionButtons} /> }
@@ -58,7 +81,6 @@ class PopupInternal extends Component {
             </div>
         );
     }
-
 }
 
 PopupInternal.propTypes = {
@@ -72,6 +94,7 @@ PopupInternal.propTypes = {
     header: PropTypes.shape({
         className: PropTypes.string,
         text: PropTypes.string,
+        closeIcon: PropTypes.bool,
         closePopup: PropTypes.func
     }),
     footer: PropTypes.shape({
@@ -135,7 +158,10 @@ Popup.propTypes = {
             onClick: PropTypes.func
         }))
     }),
-    children: PropTypes.any,
+    children: PropTypes.oneOfType([
+        PropTypes.array,
+        PropTypes.object
+    ]),
     closePopup: PropTypes.func
 };