+++ utils/validator.js
@@ -17,6 +17,16 @@ const validators = {
     },
     length: (value, length) => {
         return value.length === length;
+    },
+    regex: (value, regex) => {
+        var regexPattern = new RegExp(regex);
+        return regexPattern.test(value);
+    },
+    numbersOnly: (value, shouldBeValidated) => {
+        if (shouldBeValidated) {
+            return /^\d+$/.test(value);
+        }
+        return true;
     }
 };
 
@@ -35,6 +45,12 @@ const defaultErrorMessagingMapping = {
     },
     length: ({ input, length }) => {
         return `OTP code must be exactly ${length} characters long`;
+    },
+    numbersOnly: ({ input }) => {
+        return `Please enter only numeric characters.`;
+    },
+    regex: ({ input }) => {
+        return `Invalid field.`;
     }
 };
 
@@ -45,6 +61,13 @@ export class Validator {
     }
 
     validateInput(input, value, inputs) {
+        if (!this.config[input]) {
+            // If there is no validation for this input - simulate passing validation
+            return {
+                isValid: true,
+                error: ''
+            };
+        }
         const { validations, validateOrder } = this.config[input];
         let error = '';
 
@@ -63,7 +86,27 @@ export class Validator {
         };
     }
 
+    validateAllFlat(inputs) {
+        // if input values ARE NOT objects with key value
+        // inputs - immutable Map with 'data' and 'edited' key filled flat with the data
+        let errors = [];
+        let computedData = inputs.get('data').merge(inputs.get('edited'));
+        let keys = computedData.keySeq().toArray();
+        keys.forEach((key) => {
+            let validationResult = this.validateInput(key, computedData.get(key));
+            if (!validationResult.isValid) {
+                errors.push({
+                    field: key,
+                    error: validationResult.error
+                });
+            }
+        });
+
+        return errors;
+    }
+
     validateAll(inputs) {
+        // if input values ARE objects with key value
         let validationError = '';
         let invalidField = '';